+++ docs/html/search/all_c.js
@@ -1,5 +1,5 @@
 var searchData=
 [
-  ['timing_20stats_41',['Timing Stats',['../md__home_brian_src_AceCommon_src_timing_stats_README.html',1,'']]],
-  ['timingstats_42',['TimingStats',['../classace__common_1_1TimingStats.html',1,'ace_common::TimingStats'],['../classace__common_1_1TimingStats.html#a8ffb99931928b622e696addddf426b96',1,'ace_common::TimingStats::TimingStats()'],['../classace__common_1_1TimingStats.html#a7dfa7d8de54f34a9d3e32a4040d7884f',1,'ace_common::TimingStats::TimingStats(const TimingStats &amp;)=default']]]
+  ['timing_20stats_42',['Timing Stats',['../md__home_brian_src_AceCommon_src_timing_stats_README.html',1,'']]],
+  ['timingstats_43',['TimingStats',['../classace__common_1_1TimingStats.html',1,'ace_common::TimingStats'],['../classace__common_1_1TimingStats.html#a8ffb99931928b622e696addddf426b96',1,'ace_common::TimingStats::TimingStats()'],['../classace__common_1_1TimingStats.html#a7dfa7d8de54f34a9d3e32a4040d7884f',1,'ace_common::TimingStats::TimingStats(const TimingStats &amp;)=default']]]
 ];