+++ docs/html/search/files_1.js
@@ -1,6 +1,4 @@
 var searchData=
 [
-  ['printfto_2eh_56',['printfTo.h',['../printfTo_8h.html',1,'']]],
-  ['printpadto_2eh_57',['printPadTo.h',['../printPadTo_8h.html',1,'']]],
-  ['pstrings_2eh_58',['pstrings.h',['../pstrings_8h.html',1,'']]]
+  ['djb2_2eh_57',['djb2.h',['../djb2_8h.html',1,'']]]
 ];