+++ docs/html/search/searchdata.js
@@ -2,7 +2,7 @@ var indexSectionsWithContent =
 {
   0: "abcdfgikloprtuvw~",
   1: "fpt",
-  2: "apu",
+  2: "adpu",
   3: "bcdfgiloprtuvw~",
   4: "bkp",
   5: "aptu"