+++ src/hash/djb2.h
@@ -1,0 +1,58 @@
+/*
+MIT License
+
+Copyright (c) 2020 Brian T. Park
+
+Permission is hereby granted, free of charge, to any person obtaining a copy
+of this software and associated documentation files (the "Software"), to deal
+in the Software without restriction, including without limitation the rights
+to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+copies of the Software, and to permit persons to whom the Software is
+furnished to do so, subject to the following conditions:
+
+The above copyright notice and this permission notice shall be included in all
+copies or substantial portions of the Software.
+
+THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+SOFTWARE.
+*/
+
+/**
+ * @file djb2.h
+ *
+ * Implement the djb2 hash algorithm (http://www.cse.yorku.ca/~oz/hash.html)
+ * for normal strings and flash strings.
+ */
+
+#ifndef ACE_COMMON_HASH_DJB2_H
+#define ACE_COMMON_HASH_DJB2_H
+
+#include <stdint.h> // uint32_t
+class __FlashStringHelper;
+
+namespace ace_common {
+
+/**
+ * Implement the djb2 hash algorithm as described in
+ * https://stackoverflow.com/questions/7666509 and
+ * http://www.cse.yorku.ca/~oz/hash.html.
+ *
+ * @param s NUL terminated string, cannot be nullptr
+ */
+uint32_t hashDjb2(const char* s);
+
+/**
+ * Same as hashDjb2(const char*) but for strings stored in Arduino Flash memory.
+ *
+ * @param fs NUL terminated string stored in Flash memory, cannot be nullptr
+ */
+uint32_t hashDjb2(const __FlashStringHelper* fs);
+
+} // ace_common
+
+#endif // ACE_COMMON_HASH_DJB2_H